file decomposition
