+++ src/components/Header/Header.module.scss
@@ -1,4 +1,5 @@
-@import "../variables";
+@import "../../scss/shared/variables.scss";
+@import "../../scss/shared/mixins.scss";
 
 .header {
   border-bottom: 1px solid $gray-line;
