+++ src/components/pages/Home/Categories/Categories.module.scss
@@ -1,4 +1,5 @@
-@import "../variables";
+@import "../../../../scss/shared/variables.scss";
+@import "../../../../scss/shared/mixins.scss";
 
 .categories {
   ul {
