--- src/scss/components/_all.scss
@@ -1,5 +1,0 @@
-@import "./header";
-@import "./button";
-@import "./categories";
-@import "./sort";
-@import "./sushi-block";
